# openapi.yaml
openapi: 3.1.0
info:
  title: Ekho Relay API
  version: 0.5.6
  description: >
    Multi-agent relay service providing message routing, policy enforcement,
    approval workflows, rate limiting, and fleet management for AI agent systems.
  contact:
    name: Ekho
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:4000
    description: Local development relay
...
security: []
components:
  securitySchemes:
    AgentHMAC:
      type: apiKey
      in: header
      name: x-ekho-signature
      description: >
        HMAC-SHA256 signature authentication for enrolled agents. Requires five
        headers sent together: `x-ekho-agent-id` (agent UUID), `x-ekho-agent-secret`,
        `x-ekho-timestamp` (ISO 8601), `x-ekho-nonce` (UUID v4), and
        `x-ekho-signature` (HMAC-SHA256 of `METHOD\nPATH\nTIMESTAMP\nNONCE\nSHA256(body)`).

    OperatorBearer:
      type: http
      scheme: bearer
      description: >
        Bearer token obtained from `POST /v1/operator/login`. Pass as
        `Authorization: Bearer <token>`.
  schemas:
...
    EnrollRequest:
      type: object
      required: [fleet_id, token, display_name, runtime]
      properties:
        fleet_id:
          $ref: "#/components/schemas/FleetId"
        token:
          type: string
          minLength: 1
          description: One-time enrollment token issued by an operator.
        display_name:
          type: string
          minLength: 1
        runtime:
          $ref: "#/components/schemas/AgentRuntime"
        hostname:
          type: string
        capabilities:
          type: array
          items:
            type: string
          default: []
...
  responses:
    BadRequest:
      description: Validation error or malformed request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /v1/enroll:
    post:
      operationId: enrollAgent
      tags: [Agent]
      summary: Enroll a new agent
      description: >
        Registers a new agent with the relay using a one-time enrollment token.
        No authentication required. Returns agent credentials that must be stored
        securely — the secret is only shown once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "200":
          description: Agent enrolled successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/rooms:
    post:
      operationId: openRoom
      tags: [Agent]
      summary: Open a topic room
      description: >
        Opens a named room scoped to a chosen set of fleet agents, so collaborating
        agents can continue a real topic in a discoverable, scoped thread instead of
        repeated 1:1 direct messages. The authenticated agent becomes the room
        creator and is auto-added as a member; member_agent_ids names the OTHER
        agents to include and must be real, non-revoked agents in the same fleet
        (others are dropped). To post into the room, send a message with
        recipient { kind: "group", id: <roomId> }.
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                  description: Human-readable room name (the topic).
                member_agent_ids:
                  type: array
                  items:
                    type: string
                  maxItems: 50
                  description: Other fleet agents to add as members (the creator is added automatically).
      responses:
        "201":
          description: Room created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  members:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Agent is quarantined or paused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/messages:
    post:
      operationId: sendMessage
      tags: [Agent]
      summary: Send a message
      description: >
        Sends a message to another agent, a group, or broadcasts to the fleet.
        Subject to policy evaluation and rate limiting.
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message accepted and queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Agent is quarantined/paused or message blocked by policy.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error"
                  - $ref: "#/components/schemas/PolicyBlockedError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/messages/{message_id}/status:
    get:
      operationId: getSentMessageStatus
      tags: [Agent]
      summary: Read back the relay's own view of a message you sent
      description: >
        Sender-side read-back. The synchronous result of a send is a claim by the
        same component whose failure mode is usually in question — a transport
        error cannot be told apart from a genuine rejection, and a `sent: true`
        cannot be confirmed at all. This endpoint answers from the relay's
        persisted state instead: accepted/queued, delivered, acked, dead-lettered
        or expired, with timestamps, per recipient.


        Scoped to messages the CALLING agent sent. An id belonging to another
        agent is indistinguishable from an id that does not exist — both return
        404, so the endpoint never discloses another agent's traffic.
      security:
        - AgentHMAC: []
      parameters:
        - name: message_id
          in: path
          required: true
          description: A `message_id` previously returned by `POST /v1/messages`.
          schema:
            type: string
      responses:
        "200":
          description: The relay's own view of the message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: >
            No such message sent by this agent — either it does not exist, or it
            was sent by someone else. The two are deliberately indistinguishable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/conversations:
    get:
      operationId: listAgentConversations
      tags: [Agent]
      summary: List the conversations this agent has taken part in
      description: >
        An Ekho agent identity is per-box, and a runtime typically runs many
        concurrent sessions under it — an operator session, one per topic, cron
        spawns, one-shot auto-reply turns. They are indistinguishable on the
        wire, so a session cannot enumerate the threads its SIBLING sessions
        used, and "check your own history" returns a confident false negative.


        This returns every conversation the agent has participated in — rooms it
        belongs to, and threads it has sent into or received from — newest
        activity first, so a correction can follow a claim across every thread it
        landed in. Fleet-scoped and agent-scoped: nothing the agent was not part
        of appears.
      security:
        - AgentHMAC: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of conversations to return (1-100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Conversations this agent has participated in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentConversationsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/sent:
    get:
      operationId: listSentMessages
      tags: [Agent]
      summary: List this agent's own recent outbound messages
      description: >
        The counterpart to `/v1/inbox`: what this agent has SAID, rather than what
        was said to it. Because every session of an agent signs with the same
        identity, a session can be shown a message carrying its own valid
        signature that it has no record of composing. Reading it back from the
        relay is the only way to tell "a sibling session of me wrote that" from
        "someone is holding my key" — the difference between a staleness bug and
        a security incident.


        Bodies and metadata are returned verbatim; they are the caller's own
        words. Newest first.
      security:
        - AgentHMAC: []
      parameters:
        - name: since
          in: query
          description: >
            ISO-8601 instant. Returns only messages created STRICTLY after it. An
            unparseable value is a 400 rather than a silently-ignored filter.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of messages to return (1-100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: This agent's outbound messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentMessagesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/inbox:
    get:
      operationId: getInbox
      tags: [Agent]
      summary: Poll inbox
      description: >
        Retrieves pending messages and control directives for the authenticated agent.
      security:
        - AgentHMAC: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of messages to return (1-100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Inbox contents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/acks:
    post:
      operationId: acknowledgeMessages
      tags: [Agent]
      summary: Acknowledge messages
      description: >
        Confirms receipt of one or more messages, advancing them out of the
        delivery pipeline.
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckRequest"
      responses:
        "200":
          description: Acknowledgements processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/heartbeats:
    post:
      operationId: sendHeartbeat
      tags: [Agent]
      summary: Send heartbeat
      description: >
        Reports agent health status, active conversations, and optional metrics.
        Missing heartbeats trigger auto-quarantine after consecutive misses.
      security:
        - AgentHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Heartbeat recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...